default message

+++ 0x0B-malloc_free/1-strdup.c
@@ -11,14 +11,14 @@ char *_strdup(char *str)
 {
 	char *str2;
 
-	int len;
+	unsigned long int len;
 
 	len = 0;
 	str2 = malloc(sizeof(str));
 
 	while (str[len] != '\0')
 	{
-		str2 [len] = str[len];
+		str2[len] = str[len];
 		len++;
 	}
 	if (str == NULL || str2 == NULL)
